+++ src/affine.rs
@@ -407,7 +407,7 @@ where
     /// Returns a mutable reference to the affine transform's rows.
     #[inline]
     #[must_use]
-    pub const fn as_rows_mut(&mut self) -> &mut [Vector<2, T, A>; 3] {
+    pub const fn as_mut_rows(&mut self) -> &mut [Vector<2, T, A>; 3] {
         // SAFETY: `Affine<2, T, A>` is guaranteed to begin with
         // `Matrix<2, T, A>` (two vectors) then `Vector<2, T, A>`, which is 3
         // vectors in total.
@@ -504,7 +504,7 @@ where
     /// Returns a mutable reference to the affine transform's rows.
     #[inline]
     #[must_use]
-    pub const fn as_rows_mut(&mut self) -> &mut [Vector<3, T, A>; 4] {
+    pub const fn as_mut_rows(&mut self) -> &mut [Vector<3, T, A>; 4] {
         // SAFETY: `Affine<3, T, A>` is guaranteed to begin with
         // `Matrix<3, T, A>` (three vectors) then `Vector<3, T, A>`, which is 4
         // vectors in total.
@@ -568,7 +568,7 @@ where
     /// Returns a mutable reference to the affine transform's rows.
     #[inline]
     #[must_use]
-    pub const fn as_rows_mut(&mut self) -> &mut [Vector<4, T, A>; 5] {
+    pub const fn as_mut_rows(&mut self) -> &mut [Vector<4, T, A>; 5] {
         // SAFETY: `Affine<4, T, A>` is guaranteed to begin with
         // `Matrix<4, T, A>` (four vectors) then `Vector<4, T, A>`, which is 5
         // vectors in total.
@@ -1411,13 +1411,13 @@ mod tests {
     fn test_as_rows_mut() {
         for_types!(|T: PrimitiveNumber, A| {
             let mut rows = std::array::from_fn(|r| Vector::from_fn(|c| T::as_from(r * 2 + c)));
-            assert_eq!(Affine::<2, T, A>::from_rows(&rows).as_rows_mut(), &mut rows);
+            assert_eq!(Affine::<2, T, A>::from_rows(&rows).as_mut_rows(), &mut rows);
 
             let mut rows = std::array::from_fn(|r| Vector::from_fn(|c| T::as_from(r * 3 + c)));
-            assert_eq!(Affine::<3, T, A>::from_rows(&rows).as_rows_mut(), &mut rows);
+            assert_eq!(Affine::<3, T, A>::from_rows(&rows).as_mut_rows(), &mut rows);
 
             let mut rows = std::array::from_fn(|r| Vector::from_fn(|c| T::as_from(r * 4 + c)));
-            assert_eq!(Affine::<4, T, A>::from_rows(&rows).as_rows_mut(), &mut rows);
+            assert_eq!(Affine::<4, T, A>::from_rows(&rows).as_mut_rows(), &mut rows);
         });
     }