+++ src/quaternion.rs
@@ -229,8 +229,8 @@ where
     /// last element `w` is the real part.
     #[inline]
     #[must_use]
-    pub const fn as_array_ref(&self) -> &[T; 4] {
-        self.0.as_array_ref()
+    pub const fn as_array(&self) -> &[T; 4] {
+        self.0.as_array()
     }
 
     /// Returns a mutable reference to the quaternion's elements.
@@ -239,8 +239,8 @@ where
     /// last element `w` is the real part.
     #[inline]
     #[must_use]
-    pub const fn as_array_mut(&mut self) -> &mut [T; 4] {
-        self.0.as_array_mut()
+    pub const fn as_mut_array(&mut self) -> &mut [T; 4] {
+        self.0.as_mut_array()
     }
 
     /// Converts the quaternion `self` to a 4-dimensional vector.
@@ -257,7 +257,7 @@ where
     /// `x`, `y` and `z` are the imaginary parts and `w` is the real part.
     #[inline]
     #[must_use]
-    pub const fn as_vector_ref(&self) -> &Vector<4, T, A> {
+    pub const fn as_vector(&self) -> &Vector<4, T, A> {
         &self.0
     }
 
@@ -267,7 +267,7 @@ where
     /// `x`, `y` and `z` are the imaginary parts and `w` is the real part.
     #[inline]
     #[must_use]
-    pub const fn as_vector_mut(&mut self) -> &mut Vector<4, T, A> {
+    pub const fn as_mut_vector(&mut self) -> &mut Vector<4, T, A> {
         &mut self.0
     }
 
@@ -425,7 +425,7 @@ where
 {
     #[inline]
     fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
-        self.as_array_ref().hash(state);
+        self.as_array().hash(state);
     }
 }
 
@@ -771,7 +771,7 @@ mod tests {
             let [x, y, z, w] = std::array::from_fn(|i| T::as_from(i + 1));
 
             assert_eq!(
-                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_array_ref(),
+                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_array(),
                 &[x, y, z, w]
             );
         });
@@ -783,7 +783,7 @@ mod tests {
             let [x, y, z, w] = std::array::from_fn(|i| T::as_from(i + 1));
 
             assert_eq!(
-                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_array_mut(),
+                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_mut_array(),
                 &mut [x, y, z, w]
             );
         });
@@ -807,7 +807,7 @@ mod tests {
             let [x, y, z, w] = std::array::from_fn(|i| T::as_from(i + 1));
 
             assert_eq!(
-                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_vector_ref(),
+                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_vector(),
                 &Vector::<4, T, A>::new(x, y, z, w)
             );
         });
@@ -819,7 +819,7 @@ mod tests {
             let [x, y, z, w] = std::array::from_fn(|i| T::as_from(i + 1));
 
             assert_eq!(
-                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_vector_mut(),
+                Quaternion::<T, A>::from_xyzw(x, y, z, w).as_mut_vector(),
                 &mut Vector::<4, T, A>::new(x, y, z, w)
             );
         });